+++ src/blockchain-listener/blockchain-listener.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ViemModule } from '../viem/viem.module.js';
+import { ViemModule } from '../web3/web.module.js';
 import { BlockchainListenerService } from './blockchain-listener.service.js';
 
 export const BOTCHAIN_TOKEN = 'BotChain';
