+++ src/web3/web.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ViemService } from './viem.service.js';
+import { ViemService } from './web.service.js';
 
 @Module({
   providers: [ViemService],
